@@ -1,31 +1,24 @@
 {
    "name": "snake",
    "version": "1.0.0",
-   "description": "game the snake",
-   "private": true,
-   "author": "florian.bergot564@gmail.com",
+   "main": "./src/app.js",
+   "type": "module",
+   "author": "fbergot <florian.bergot564@gmail.com>",
    "license": "MIT",
    "scripts": {
       "build_w": "webpack --watch",
       "build": "webpack",
       "test": "jest --watch"
    },
-   "dependencies": {
-      "prettier": "^2.5.1",
-      "webpack": "^5.68.0"
-   },
    "devDependencies": {
-      "babel-core": "6",
-      "babel-jest": "^28.1.1",
-      "babel-loader": "7",
-      "babel-preset-env": "^1.7.0",
-      "css-loader": "^6.6.0",
-      "eslint": "^8.17.0",
+      "@babel/core": "^7.17.8",
+      "@babel/preset-env": "^7.16.11",
+      "babel-loader": "^8.2.4",
+      "css-loader": "^6.7.1",
       "file-loader": "^6.2.0",
-      "jest": "^28.1.1",
-      "jest-environment-jsdom": "^28.1.1",
+      "prettier": "^2.6.1",
       "style-loader": "^3.3.1",
-      "webpack-cli": "^4.9.2",
-      "webpack-dev-server": "^4.7.4"
+      "webpack": "^5.70.0",
+      "webpack-cli": "^4.9.2"
    }
 }
