@@ -1,8 +1,8 @@
 import "./assets/CSS/style.css";
-import Game from "./class/Game";
-import { images } from "./function/imgBuilder";
-import { sounds } from "./function/soundBuilder";
-import GameAssetsDecorator from "./function/decorator/game_assets_dec";
+import Game from "./class/Game.js";
+import { images } from "./function/imgBuilder.js";
+import { sounds } from "./function/soundBuilder.js";
+import GameAssetsDecorator from "./function/decorator/game_assets_dec.js";
 
 const game = new (GameAssetsDecorator(Game, { images, sounds }))();
 game.start();
