@@ -1,4 +1,4 @@
-import Utils from "./Utils";
+import Utils from "./Utils.js";
 
 class SpeedManager {
    constructor(nForTrigger, speedsLoopObject) {
